% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{merge3}
\alias{merge3}
\title{create three way merges}
\usage{
merge3(base, merge1, merge2, level = "minimal", favor = NULL, style = NULL)
}
\arguments{
\item{base}{base file}

\item{merge1, merge2}{file with changes}

\item{level}{"minimal", "eager", "zealous", "zealous_alnum"}

\item{favor}{merge favor, "ours", "theirs", or "union"}

\item{style}{merge style, "diff3", or "zdiff3"}
}
\description{
create three way merges
}
