% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unidiff.R
\name{unidiff}
\alias{unidiff}
\title{create unified diffs}
\usage{
unidiff(old, new, create_head = TRUE)
}
\arguments{
\item{old}{oldfile}

\item{new}{oldfile}

\item{create_head}{add a header to the file}
}
\description{
create unified diffs
}
