% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unidiff_dir.R
\name{unidiff_dir}
\alias{unidiff_dir}
\title{Function to diff files between two directories}
\usage{
unidiff_dir(old, new)
}
\arguments{
\item{old, new}{directories to check}
}
\description{
Function to diff files between two directories
}
