% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unidiff_dir.R
\name{unidiff_dir}
\alias{unidiff_dir}
\title{Function to diff files between two directories}
\usage{
unidiff_dir(old, new, pattern = NULL)
}
\arguments{
\item{old, new}{directories to check}

\item{pattern}{an optional \link[base]{regular expression}.  Only file names
    which match the regular expression will be returned.}
}
\description{
Function to diff files between two directories
}
